stopped favourite applicants from being automatically deleted from the database after 1 year

+++ Group Project/src/main/java/org/example/groupproject/applicant/ApplicantScheduler.java
@@ -17,7 +17,7 @@ public class ApplicantScheduler {
     public void deleteApplicantsInDatabaseForMoreThanAYear() {
         List<Applicant> applicantList = applicantRepository.findAll();
         for (Applicant applicant : applicantList) {
-            if (applicantRepository.hasBeenInDatabaseForMoreThanAYear(applicant.id().intValue())) {
+            if (applicantRepository.hasBeenInDatabaseForMoreThanAYear(applicant.id().intValue()) && !applicant.isFavourite()) {
                 applicantRepository.deleteById(applicant.id().intValue());
             }
         }
